Refactor shared module
