+++ imessage-exporter/src/exporters/html/balloons.rs
@@ -16,7 +16,7 @@ use askama::Template;
 
 use crate::exporters::{
     exporter::{BalloonFormatter, MessageFormatter},
-    shared::format::format_check_in_caption,
+    shared::balloon::format_check_in_caption,
 };
 
 use super::{
