+++ imessage-exporter/src/exporters/shared/mod.rs
@@ -2,4 +2,4 @@ pub(crate) mod announcement;
 pub(crate) mod balloon;
 pub(crate) mod driver;
 pub(crate) mod edited;
-pub(crate) mod format;
+pub(crate) mod time;
